test: Migrate tests to typescript

+++ test/functional/simulator-e2e-specs.ts
@@ -5,19 +5,21 @@ import { getSimulator } from '../../lib/simulator';
 import { Simctl } from 'node-simctl';
 import B from 'bluebird';
 import { retryInterval, waitForCondition } from 'asyncbox';
-import path from 'path';
 import xcode from 'appium-xcode';
 import { LONG_TIMEOUT, verifyStates } from './helpers';
+import { use as chaiUse, expect } from 'chai';
+import chaiAsPromised from 'chai-as-promised';
+import { getUIKitCatalogPath, UICATALOG_BUNDLE_ID } from '../setup';
 
-const BUNDLE_ID = 'io.appium.TestApp';
-const OS_VERSION = process.env.MOBILE_OS_VERSION || '16.2';
-const DEVICE_NAME = process.env.MOBILE_DEVICE_NAME || 'iPhone 14';
-const CUSTOM_APP = path.resolve(__dirname, '..', 'assets', 'TestApp-iphonesimulator.app');
+chaiUse(chaiAsPromised);
 
-async function deleteSimulator (udid, version) {
+const OS_VERSION = process.env.MOBILE_OS_VERSION || '26.0';
+const DEVICE_NAME = process.env.MOBILE_DEVICE_NAME || 'iPhone 17';
+
+async function deleteSimulator (udid: string, version: string): Promise<void> {
   // only want to get rid of the device if it is present
   const simctl = new Simctl();
-  let devices = await simctl.getDevices();
+  const devices = await simctl.getDevices();
   if (!devices[version]) {
     return;
   }
@@ -35,18 +37,13 @@ describe(`simulator ${OS_VERSION}`, function () {
   this.timeout(LONG_TIMEOUT);
   this.retries(2);
 
-  let simctl;
-  let chai;
-  let xcodeVersion;
+  let simctl: Simctl;
+  let xcodeVersion: any;
+  let customApp: string;
 
   before(async function () {
-    chai = await import('chai');
-    const chaiAsPromised = await import('chai-as-promised');
-
-    chai.should();
-    chai.use(chaiAsPromised.default);
-
     xcodeVersion = await xcode.getVersion(true);
+    customApp = await getUIKitCatalogPath();
   });
 
   beforeEach(async function () {
@@ -61,94 +58,117 @@ describe(`simulator ${OS_VERSION}`, function () {
   });
   afterEach(async function () {
     await killAllSimulators();
-    await deleteSimulator(simctl.udid, OS_VERSION);
+    if (simctl.udid) {
+      await deleteSimulator(simctl.udid, OS_VERSION);
+    }
   });
 
   it('should detect whether a simulator has been run before', async function () {
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
     const sim = await getSimulator(simctl.udid);
-    await sim.isFresh().should.eventually.be.true;
+    await expect(sim.isFresh()).to.eventually.be.true;
     await sim.run({startupTimeout: LONG_TIMEOUT / 2});
-    await sim.isFresh().should.eventually.be.false;
+    await expect(sim.isFresh()).to.eventually.be.false;
     await sim.shutdown();
     await sim.clean();
-    await sim.isFresh().should.eventually.be.true;
+    await expect(sim.isFresh()).to.eventually.be.true;
   });
 
   it('should launch and shutdown a sim', async function () {
-    let sim = await getSimulator(simctl.udid);
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
+    const sim = await getSimulator(simctl.udid);
     await sim.run({startupTimeout: LONG_TIMEOUT / 2});
     await sim.shutdown();
-    (await sim.stat()).state.should.equal('Shutdown');
+    expect((await sim.stat()).state).to.equal('Shutdown');
   });
 
   it('should be able to delete an app', async function () {
-    let sim = await getSimulator(simctl.udid);
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
+    const sim = await getSimulator(simctl.udid);
     await sim.run({startupTimeout: LONG_TIMEOUT});
 
     // install & launch test app
-    await sim.installApp(CUSTOM_APP);
+    await sim.installApp(customApp);
 
     console.log('Application installed'); // eslint-disable-line no-console
 
-    await sim.isAppInstalled(BUNDLE_ID).should.eventually.be.true;
+    await expect(sim.isAppInstalled(UICATALOG_BUNDLE_ID)).to.eventually.be.true;
 
     // this remains somewhat flakey
     await retryInterval(5, 1000, async () => {
-      await sim.launchApp(BUNDLE_ID, 1);
+      await sim.launchApp(UICATALOG_BUNDLE_ID, {wait: true});
     });
 
     console.log('Application launched'); // eslint-disable-line no-console
 
     // Wait for application process
     await waitForCondition(
-      async () => (await sim.ps()).some(({name}) => name === BUNDLE_ID), {
+      async () => (await sim.ps()).some(({name}) => name === UICATALOG_BUNDLE_ID), {
         waitMs: 10000,
         intervalMs: 500,
       });
 
-    await sim.removeApp(BUNDLE_ID);
+    await sim.removeApp(UICATALOG_BUNDLE_ID);
 
     // should not be able to launch anymore
-    await sim.launchApp(BUNDLE_ID, 1).should.eventually.be.rejected;
+    await expect(sim.launchApp(UICATALOG_BUNDLE_ID, {wait: true})).to.eventually.be.rejected;
 
-    (await sim.isAppInstalled(BUNDLE_ID)).should.be.false;
+    expect(await sim.isAppInstalled(UICATALOG_BUNDLE_ID)).to.be.false;
   });
 
   it('should delete a sim', async function () {
-    let sim = await getSimulator(simctl.udid);
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
+    const sim = await getSimulator(simctl.udid);
     await sim.delete();
-    await getSimulator(simctl.udid).should.eventually.be.rejected;
+    await expect(getSimulator(simctl.udid)).to.eventually.be.rejected;
   });
 
   it('should start a sim using the "run" method', async function () {
-    let sim = await getSimulator(simctl.udid);
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
+    const sim = await getSimulator(simctl.udid);
 
     await sim.run({startupTimeout: LONG_TIMEOUT});
 
-    (await sim.stat()).state.should.equal('Booted');
+    expect((await sim.stat()).state).to.equal('Booted');
 
     await sim.shutdown();
-    (await sim.stat()).state.should.equal('Shutdown');
+    expect((await sim.stat()).state).to.equal('Shutdown');
   });
 
   it('should be able to start safari', async function () {
-    let sim = await getSimulator(simctl.udid);
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
+    const sim = await getSimulator(simctl.udid);
 
     await sim.run({startupTimeout: LONG_TIMEOUT});
     await sim.openUrl('https://apple.com');
-    await sim.isAppRunning(MOBILE_SAFARI_BUNDLE_ID).should.eventually.be.true;
+    await expect(sim.isAppRunning(MOBILE_SAFARI_BUNDLE_ID)).to.eventually.be.true;
     await sim.shutdown();
   });
 
   it('should detect if a sim is running', async function () {
-    let sim = await getSimulator(simctl.udid);
-    await sim.isRunning().should.eventually.be.false;
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
+    const sim = await getSimulator(simctl.udid);
+    await expect(sim.isRunning()).to.eventually.be.false;
 
     await sim.run({startupTimeout: LONG_TIMEOUT});
-    await sim.isRunning().should.eventually.be.true;
+    await expect(sim.isRunning()).to.eventually.be.true;
 
     await sim.shutdown();
-    await sim.isRunning().should.eventually.be.false;
+    await expect(sim.isRunning()).to.eventually.be.false;
   });
 
   it('should properly start simulator in headless mode on Xcode9+', async function () {
@@ -156,6 +176,9 @@ describe(`simulator ${OS_VERSION}`, function () {
       return this.skip();
     }
 
+    if (!simctl.udid) {
+      throw new Error('simctl.udid is null');
+    }
     const sim = await getSimulator(simctl.udid);
     await verifyStates(sim, false, false);
 
@@ -180,16 +203,9 @@ describe(`reuse an already-created already-run simulator ${OS_VERSION}`, functio
   this.timeout(LONG_TIMEOUT);
   this.retries(2);
 
-  let sim;
-  let chai;
+  let sim: any;
 
   before(async function () {
-    chai = await import('chai');
-    const chaiAsPromised = await import('chai-as-promised');
-
-    chai.should();
-    chai.use(chaiAsPromised.default);
-
     await killAllSimulators();
     const udid = await new Simctl().createDevice(
       'ios-simulator testing',
@@ -208,28 +224,23 @@ describe(`reuse an already-created already-run simulator ${OS_VERSION}`, functio
   it('should start a sim using the "run" method', async function () {
     await sim.run({startupTimeout: LONG_TIMEOUT});
 
-    (await sim.stat()).state.should.equal('Booted');
+    expect((await sim.stat()).state).to.equal('Booted');
 
     await sim.shutdown();
-    (await sim.stat()).state.should.equal('Shutdown');
+    expect((await sim.stat()).state).to.equal('Shutdown');
   });
 });
 
 describe('advanced features', function () {
-  let sim;
-  let chai;
-  let xcodeVersion;
+  let sim: any;
+  let xcodeVersion: any;
+  let customApp: string;
 
   this.timeout(LONG_TIMEOUT);
 
   before(async function () {
-    chai = await import('chai');
-    const chaiAsPromised = await import('chai-as-promised');
-
-    chai.should();
-    chai.use(chaiAsPromised.default);
-
     xcodeVersion = await xcode.getVersion(true);
+    customApp = await getUIKitCatalogPath();
 
     await killAllSimulators();
     const udid = await new Simctl().createDevice(
@@ -247,28 +258,28 @@ describe('advanced features', function () {
   });
 
   describe('custom apps', function () {
-    it('should find bundle id for TestApp', async function () {
-      if (!await sim.isAppInstalled(CUSTOM_APP)) {
-        await sim.installApp(CUSTOM_APP);
+    it('should find bundle id for UIKitCatalog', async function () {
+      if (!await sim.isAppInstalled(customApp)) {
+        await sim.installApp(customApp);
       }
-      if (!await sim.isAppRunning(CUSTOM_APP)) {
-        await sim.launchApp(BUNDLE_ID);
+      if (!await sim.isAppRunning(customApp)) {
+        await sim.launchApp(UICATALOG_BUNDLE_ID);
       }
 
-      await sim.getUserInstalledBundleIdsByBundleName('TestApp').should.eventually.eql([BUNDLE_ID]);
+      await expect(sim.getUserInstalledBundleIdsByBundleName('UIKitCatalog')).to.eventually.eql([UICATALOG_BUNDLE_ID]);
     });
 
     it('should scrub custom app', async function () {
-      if (!await sim.isAppInstalled(CUSTOM_APP)) {
-        await sim.installApp(CUSTOM_APP);
+      if (!await sim.isAppInstalled(customApp)) {
+        await sim.installApp(customApp);
       }
-      if (!await sim.isAppRunning(CUSTOM_APP)) {
-        await sim.launchApp(BUNDLE_ID);
+      if (!await sim.isAppRunning(customApp)) {
+        await sim.launchApp(UICATALOG_BUNDLE_ID);
       }
-      await sim.scrubApp(BUNDLE_ID);
-      await sim.isAppRunning(BUNDLE_ID).should.eventually.be.false;
-      await sim.launchApp(BUNDLE_ID);
-      await sim.isAppRunning(BUNDLE_ID).should.eventually.be.true;
+      await sim.scrubApp(UICATALOG_BUNDLE_ID);
+      await expect(sim.isAppRunning(UICATALOG_BUNDLE_ID)).to.eventually.be.false;
+      await sim.launchApp(UICATALOG_BUNDLE_ID);
+      await expect(sim.isAppRunning(UICATALOG_BUNDLE_ID)).to.eventually.be.true;
     });
   });
 
@@ -278,7 +289,7 @@ describe('advanced features', function () {
         return this.skip();
       }
       await sim.enrollBiometric(true);
-      (await sim.isBiometricEnrolled()).should.be.true;
+      expect(await sim.isBiometricEnrolled()).to.be.true;
     });
 
     it(`should properly enroll biometric to disabled state`, async function () {
@@ -286,7 +297,7 @@ describe('advanced features', function () {
         return this.skip();
       }
       await sim.enrollBiometric(false);
-      (await sim.isBiometricEnrolled()).should.be.false;
+      expect(await sim.isBiometricEnrolled()).to.be.false;
     });
   });
 
@@ -308,19 +319,19 @@ describe('advanced features', function () {
           name: 'en_US',
           layout: 'QWERTY',
         }
-      })).should.be.true;
+      })).to.be.true;
     });
   });
 
   describe('keychains', function () {
     it('should properly backup and restore Simulator keychains', async function () {
       if (await sim.backupKeychains()) {
-        (await sim.restoreKeychains('*.db*')).should.be.true;
+        expect(await sim.restoreKeychains('*.db*')).to.be.true;
       }
     });
 
     it('should clear Simulator keychains while it is running', async function () {
-      await sim.clearKeychains().should.eventually.be.fulfilled;
+      await expect(sim.clearKeychains()).to.eventually.be.fulfilled;
     });
   });
 
@@ -353,9 +364,9 @@ describe('advanced features', function () {
       }
       await sim.launchApp(MOBILE_SAFARI_BUNDLE_ID, {wait: true});
       await sim.scrubSafari();
-      await sim.isAppRunning(MOBILE_SAFARI_BUNDLE_ID).should.eventually.be.false;
+      await expect(sim.isAppRunning(MOBILE_SAFARI_BUNDLE_ID)).to.eventually.be.false;
       await sim.launchApp(MOBILE_SAFARI_BUNDLE_ID, {wait: true});
-      await sim.isAppRunning(MOBILE_SAFARI_BUNDLE_ID).should.eventually.be.true;
+      await expect(sim.isAppRunning(MOBILE_SAFARI_BUNDLE_ID)).to.eventually.be.true;
     });
 
     it('should set arbitrary preferences on Safari', async function () {
@@ -369,23 +380,23 @@ describe('advanced features', function () {
   describe('Permission', function () {
     it('should set and get with simctrl privacy command', async function () {
       // no exceptions
-      await chai.expect(sim.setPermission('com.apple.Maps', 'location', 'yes')).not.to.be.rejected;
-      await chai.expect(sim.setPermission('com.apple.Maps', 'location', 'NO')).not.to.be.rejected;
-      await chai.expect(sim.setPermission('com.apple.Maps', 'location', 'unset')).not.to.be.rejected;
-      await chai.expect(sim.setPermission('com.apple.Maps', 'location', 'unsupported')).to.be.rejected;
+      await expect(sim.setPermission('com.apple.Maps', 'location', 'yes')).to.not.be.rejected;
+      await expect(sim.setPermission('com.apple.Maps', 'location', 'NO')).to.not.be.rejected;
+      await expect(sim.setPermission('com.apple.Maps', 'location', 'unset')).to.not.be.rejected;
+      await expect(sim.setPermission('com.apple.Maps', 'location', 'unsupported')).to.be.rejected;
     });
 
     it('should set and get with wix command', async function () {
       await sim.setPermission('com.apple.Maps', 'contacts', 'yes');
-      await sim.getPermission('com.apple.Maps', 'contacts').should.eventually.eql('yes');
+      await expect(sim.getPermission('com.apple.Maps', 'contacts')).to.eventually.eql('yes');
       await sim.setPermission('com.apple.Maps', 'contacts', 'no');
-      await sim.getPermission('com.apple.Maps', 'contacts').should.eventually.eql('no');
+      await expect(sim.getPermission('com.apple.Maps', 'contacts')).to.eventually.eql('no');
 
       // unset sets as 'no'
       await sim.setPermission('com.apple.Maps', 'contacts', 'yes');
-      await sim.getPermission('com.apple.Maps', 'contacts').should.eventually.eql('yes');
+      await expect(sim.getPermission('com.apple.Maps', 'contacts')).to.eventually.eql('yes');
       await sim.setPermission('com.apple.Maps', 'contacts', 'unset');
-      await sim.getPermission('com.apple.Maps', 'contacts').should.eventually.eql('no');
+      await expect(sim.getPermission('com.apple.Maps', 'contacts')).to.eventually.eql('no');
     });
   });
 });
@@ -394,17 +405,11 @@ describe(`multiple instances of ${OS_VERSION} simulator on Xcode9+`, function ()
   this.timeout(LONG_TIMEOUT * 2);
   this.retries(2);
 
-  let simulatorsMapping = {};
-  let chai;
-  let xcodeVersion;
+  let simulatorsMapping: Record<string, any> = {};
+  let xcodeVersion: any;
   const DEVICES_COUNT = 2;
 
   before(async function () {
-    chai = await import('chai');
-    const chaiAsPromised = await import('chai-as-promised');
-
-    chai.should();
-    chai.use(chaiAsPromised.default);
 
     xcodeVersion = await xcode.getVersion(true);
     if (xcodeVersion.major < 9) {
@@ -430,7 +435,7 @@ describe(`multiple instances of ${OS_VERSION} simulator on Xcode9+`, function ()
         try {
           simctl.udid = udid;
           await simctl.deleteDevice();
-        } catch (err) {
+        } catch (err: any) {
           console.log(`Error deleting simulator '${udid}': ${err.message}`); // eslint-disable-line
         }
       }
@@ -450,7 +455,7 @@ describe(`multiple instances of ${OS_VERSION} simulator on Xcode9+`, function ()
     });
 
     // Should be called before launching simulator
-    await simulators[0].getUserInstalledBundleIdsByBundleName('UICatalog').should.eventually.eql([]);
+    await expect(simulators[0].getUserInstalledBundleIdsByBundleName('UICatalog')).to.eventually.eql([]);
 
     for (const sim of _.values(simulatorsMapping)) {
       await sim.run({startupTimeout: LONG_TIMEOUT});
@@ -470,16 +475,9 @@ describe(`multiple instances of ${OS_VERSION} simulator on Xcode9+`, function ()
 
 describe('getWebInspectorSocket', function () {
   this.timeout(LONG_TIMEOUT);
-  let sim;
-  let chai;
+  let sim: any;
 
   before(async function () {
-    chai = await import('chai');
-    const chaiAsPromised = await import('chai-as-promised');
-
-    chai.should();
-    chai.use(chaiAsPromised.default);
-
     await killAllSimulators();
     const udid = await new Simctl().createDevice(
       'ios-simulator testing',
@@ -495,17 +493,17 @@ describe('getWebInspectorSocket', function () {
     await deleteSimulator(sim.udid, OS_VERSION);
   });
   it('should get a socket when appropriate', async function () {
-    let socket = await sim.getWebInspectorSocket();
+    const socket = await sim.getWebInspectorSocket();
 
     if (parseFloat(OS_VERSION) < 11.3) {
-      chai.expect(socket).to.be.null;
+      expect(socket).to.be.null;
     } else {
-      socket.should.include('/private/tmp/com.apple.launchd');
-      socket.should.include('com.apple.webinspectord_sim.socket');
+      expect(socket).to.include('/private/tmp/com.apple.launchd');
+      expect(socket).to.include('com.apple.webinspectord_sim.socket');
     }
   });
   describe('two simulators', function () {
-    let sim2;
+    let sim2: any;
 
     before(async function () {
       if (parseFloat(OS_VERSION) < 11.3) {
@@ -524,27 +522,28 @@ describe('getWebInspectorSocket', function () {
     });
     after(async function () {
       await killAllSimulators();
-      if (sim2) {
+      if (sim2 && sim2.udid) {
         await deleteSimulator(sim2.udid, OS_VERSION);
       }
     });
     it('should not confuse two different simulators', async function () {
-      let socket = await sim.getWebInspectorSocket();
-      socket.should.exist;
+      const socket = await sim.getWebInspectorSocket();
+      expect(socket).to.exist;
 
-      let socket2 = await sim2.getWebInspectorSocket();
-      socket2.should.exist;
+      const socket2 = await sim2.getWebInspectorSocket();
+      expect(socket2).to.exist;
 
-      socket.should.not.eql(socket2);
+      expect(socket).to.not.eql(socket2);
     });
     it('should always get the same socket', async function () {
       let socket = await sim.getWebInspectorSocket();
       for (let i = 0; i < 10; i++) {
         sim.webInspectorSocket = null;
-        let socket2 = await sim.getWebInspectorSocket();
-        socket.should.eql(socket2);
+        const socket2 = await sim.getWebInspectorSocket();
+        expect(socket).to.eql(socket2);
         socket = socket2;
       }
     });
   });
 });
+
